Some minor cleanup. You can write a float literal as 1.23f; no need to cast. Also, I adjusted the knockback snowball so that it uses more finals; it's a good practice not to put different things in the same field or variable, and I think it is worth doing a few extra sqrt operations.

+++ SnowballMadness/src/snowballmadness/SnowballLogic.java
@@ -232,10 +232,10 @@ public abstract class SnowballLogic {
                 return new LinkedWaterTrailSnowballLogic(Material.WATER_LILY);
 
             case TNT:
-                return new TNTSnowballLogic(4);
+                return new TNTSnowballLogic(4.0f);
 
             case SULPHUR:
-                return new TNTSnowballLogic((float)1.4);
+                return new TNTSnowballLogic(1.4f);
 
             case FIREWORK:
                 return new JetpackSnowballLogic();
